+++ purewiki/core/config.php
@@ -68,6 +68,7 @@ function getGlobalConfig(bool $forceReload = false): array {
         'editor_show_snippet' => true,
         'editor_show_accordion' => true,
         'editor_show_grid' => true,
+        'editor_show_math' => true,
         'dev_debug_output'        => false,
         'dev_debug_log_max_size'  => 2,
         'dev_debug_log_max_files' => 3,